refactor(bundle): move template bundler to a seperate module
